requirements.txt
torch
torchvision
pandas
matplotlib
seaborn
split-folders
pillow
tqdm
python-dotenv
mlflow
onnx
onnxruntime